Update build system to use vanilla GNU toolchain
Update the archive file containing modified versions of the C Runtime
and syscalls to versions built with the vanilla GNU compiler
rather than the PULP extended compiler. This allows for the compilation
of the example programs using the vanilla GNU toolchain.

The team was informed by Embecosm that they would prefer this project be
completed using the vanilla GNU compiler as the PULP version was
outdated and sometimes unreliable.