Add call to time for random number generator seed
Replace the fixed number in the random number generator in the bubble
sort program with a call to time(NULL). On a system with an RTC/OS this
would return the current time in seconds, thus it can be used as the
seed for a pseudo-random number generator.

When using the PULP toolchain this call caused the linker to try and
link the default syscalls implementation on top of the edited one
provided in the libcv-verif.a file. Using the vanilla toolchain removed
this problem.

Since this program is running bare-metal with no RTC the call to time
always returns -1. This means that, functionally, it is no different
from providing a constant number. However, it does present a more
realistic C program, and so the work needed to port it should be more
realistic.